Fix usages to allow the backend to return the websocket url

+++ cslabs-backend/Proxmox/ProxmoxManager.cs
@@ -37,7 +37,7 @@ namespace CSLabsBackend.Proxmox
             }
 
             list = list.Where(p => p.Key.MemoryUsage.Free > requiredMemoryBytes).ToList();
-            list.Sort((s1,s2) => s1.Key.CpuUsage - s2.Key.CpuUsage);
+            list.Sort((s1,s2) => (int)(s1.Key.CpuUsage - s2.Key.CpuUsage));
             if(list.Count == 0)
                 throw new NoHypervisorAvailableException();
